fixed name component, style, fixed code
